requirements.txt
flask==3.1.2
python-dotenv==1.2.1
opencv-python-headless==4.11.0.86
pillow==12.1.1
numpy<2.0.0
paddlepaddle==2.6.2
paddleocr==2.6.1.3
easyocr==1.7.2
pytesseract==0.3.13
pdf2image==1.17.0
# Core dependencies for the above
requests
intel-openmp
intel-openmp

